Refactor desktop app ownership
