+++ erun-ui/state_handlers.go
@@ -1,0 +1,222 @@
+package main
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"os/exec"
+	"strings"
+
+	eruncommon "github.com/sophium/erun/erun-common"
+)
+
+func (a *App) LoadState() (uiState, error) {
+	result, err := eruncommon.ResolveListResult(a.deps.store, a.deps.findProjectRoot, eruncommon.OpenParams{
+		UseDefaultTenant:      true,
+		UseDefaultEnvironment: true,
+	})
+	if err != nil {
+		if errors.Is(err, eruncommon.ErrNotInitialized) {
+			info := a.deps.resolveBuildInfo()
+			return uiState{
+				Message:            "ERun is not initialized yet. Run `erun init` first.",
+				Build:              buildDetailsFrom(info),
+				VersionSuggestions: a.runtimeVersionSuggestions(info, ""),
+			}, nil
+		}
+		return uiState{}, err
+	}
+	info := a.deps.resolveBuildInfo()
+	state := stateFromListResult(result, info)
+	suggestionTenant := ""
+	if state.Selected != nil {
+		suggestionTenant = state.Selected.Tenant
+	} else if len(state.Tenants) > 0 {
+		suggestionTenant = state.Tenants[0].Name
+	}
+	state.VersionSuggestions = a.runtimeVersionSuggestions(info, suggestionTenant)
+	return state, nil
+}
+
+func (a *App) resolveRuntimeRegistryVersionsForTenant(tenant string) eruncommon.RuntimeRegistryVersions {
+	ctx := a.ctx
+	if ctx == nil {
+		ctx = context.Background()
+	}
+	repository := eruncommon.DefaultRuntimeImageName
+	if tenant = strings.TrimSpace(tenant); tenant != "" {
+		repository = eruncommon.RuntimeReleaseName(tenant)
+	}
+	versions, err := a.deps.resolveImageRegistry(ctx, eruncommon.DefaultContainerRegistry, repository)
+	if err != nil {
+		return eruncommon.RuntimeRegistryVersions{}
+	}
+	return versions
+}
+
+func (a *App) runtimeVersionSuggestions(info eruncommon.BuildInfo, tenant string) []uiVersion {
+	tenant = strings.TrimSpace(tenant)
+	if tenant == "" {
+		return labelRuntimeVersionSuggestions("ERun", eruncommon.DefaultRuntimeImageName, eruncommon.RuntimeDeployVersionSuggestions(info, a.resolveRuntimeRegistryVersionsForTenant("")))
+	}
+
+	suggestions := make([]uiVersion, 0, 8)
+	tenantImage := eruncommon.RuntimeReleaseName(tenant)
+	suggestions = append(suggestions, labelRuntimeVersionSuggestions(tenant, tenantImage, eruncommon.RuntimeDeployVersionSuggestions(info, a.resolveRuntimeRegistryVersionsForTenant(tenant)))...)
+	if tenantImage == eruncommon.DefaultRuntimeImageName {
+		return suggestions
+	}
+	suggestions = append(suggestions, labelRuntimeVersionSuggestions("ERun", eruncommon.DefaultRuntimeImageName, eruncommon.RuntimeDeployVersionSuggestions(info, a.resolveRuntimeRegistryVersionsForTenant("")))...)
+	return suggestions
+}
+
+func (a *App) LoadVersionSuggestions(selection uiSelection) ([]uiVersion, error) {
+	selection = normalizeSelection(selection)
+	if selection.Action == "init" {
+		return a.runtimeVersionSuggestions(a.deps.resolveBuildInfo(), selection.Tenant), nil
+	}
+	return a.runtimeVersionSuggestions(a.deps.resolveBuildInfo(), selection.Tenant), nil
+}
+
+func (a *App) LoadKubernetesContexts() ([]string, error) {
+	contexts, err := a.deps.listKubeContexts()
+	if err != nil {
+		return nil, err
+	}
+	return normalizeKubernetesContexts(contexts), nil
+}
+
+func labelRuntimeVersionSuggestions(source, image string, suggestions []uiVersion) []uiVersion {
+	source = strings.TrimSpace(source)
+	image = strings.TrimSpace(image)
+	labeled := make([]uiVersion, 0, len(suggestions))
+	for _, suggestion := range suggestions {
+		label := strings.TrimSpace(suggestion.Label)
+		if source != "" && label != "" {
+			label = source + " " + strings.ToLower(label[:1]) + label[1:]
+		}
+		suggestion.Label = label
+		suggestion.Source = source
+		suggestion.Image = image
+		labeled = append(labeled, suggestion)
+	}
+	return labeled
+}
+
+func stateFromListResult(result eruncommon.ListResult, info eruncommon.BuildInfo) uiState {
+	state := uiState{
+		Tenants: make([]uiTenant, 0, len(result.Tenants)),
+		Build:   buildDetailsFrom(info),
+	}
+	for _, tenant := range result.Tenants {
+		if len(tenant.Environments) == 0 {
+			continue
+		}
+		item := uiTenant{
+			Name:         strings.TrimSpace(tenant.Name),
+			Environments: make([]uiEnvironment, 0, len(tenant.Environments)),
+		}
+		for _, environment := range tenant.Environments {
+			item.Environments = append(item.Environments, uiEnvironment{
+				Name:           strings.TrimSpace(environment.Name),
+				MCPURL:         mcpEndpointForListEnvironment(environment),
+				RuntimeVersion: strings.TrimSpace(environment.RuntimeVersion),
+				IsActive:       environment.IsActive,
+				SSHDEnabled:    environment.SSH.Enabled,
+			})
+		}
+		state.Tenants = append(state.Tenants, item)
+	}
+	if result.CurrentDirectory.Effective != nil {
+		state.Selected = &uiSelection{
+			Tenant:      strings.TrimSpace(result.CurrentDirectory.Effective.Tenant),
+			Environment: strings.TrimSpace(result.CurrentDirectory.Effective.Environment),
+		}
+	}
+	return state
+}
+
+func mcpEndpointForOpenResult(result eruncommon.OpenResult) string {
+	return fmt.Sprintf("http://127.0.0.1:%d/mcp", eruncommon.MCPPortForResult(result))
+}
+
+func mcpEndpointForListEnvironment(environment eruncommon.ListEnvironmentResult) string {
+	port := environment.LocalPorts.MCP
+	if port <= 0 {
+		return ""
+	}
+	return fmt.Sprintf("http://127.0.0.1:%d/mcp", port)
+}
+
+func buildDetailsFrom(info eruncommon.BuildInfo) uiBuildDetails {
+	return uiBuildDetails{
+		Version: info.Version,
+		Commit:  info.Commit,
+		Date:    info.Date,
+	}
+}
+
+func listKubernetesContexts() ([]string, error) {
+	output, err := exec.Command("kubectl", "config", "get-contexts", "-o=name").Output()
+	if err != nil {
+		return nil, err
+	}
+	contexts := strings.Split(string(output), "\n")
+
+	currentOutput, err := exec.Command("kubectl", "config", "current-context").Output()
+	if err == nil {
+		contexts = preferCurrentKubernetesContext(contexts, string(currentOutput))
+	}
+
+	return contexts, nil
+}
+
+func normalizeKubernetesContexts(contexts []string) []string {
+	seen := make(map[string]struct{}, len(contexts))
+	result := make([]string, 0, len(contexts))
+	for _, context := range contexts {
+		context = strings.TrimSpace(context)
+		if context == "" {
+			continue
+		}
+		if _, ok := seen[context]; ok {
+			continue
+		}
+		seen[context] = struct{}{}
+		result = append(result, context)
+	}
+	return result
+}
+
+func preferCurrentKubernetesContext(contexts []string, current string) []string {
+	current = strings.TrimSpace(current)
+	if current == "" {
+		return contexts
+	}
+
+	result := make([]string, 0, len(contexts)+1)
+	result = append(result, current)
+	for _, context := range contexts {
+		if strings.TrimSpace(context) == current {
+			continue
+		}
+		result = append(result, context)
+	}
+	return result
+}
+
+func normalizeSelection(selection uiSelection) uiSelection {
+	return uiSelection{
+		Tenant:            strings.TrimSpace(selection.Tenant),
+		Environment:       strings.TrimSpace(selection.Environment),
+		Version:           strings.TrimSpace(selection.Version),
+		RuntimeImage:      strings.TrimSpace(selection.RuntimeImage),
+		KubernetesContext: strings.TrimSpace(selection.KubernetesContext),
+		ContainerRegistry: strings.TrimSpace(selection.ContainerRegistry),
+		NoGit:             selection.NoGit,
+		Bootstrap:         selection.Bootstrap,
+		SetDefaultTenant:  selection.SetDefaultTenant,
+		Action:            strings.TrimSpace(selection.Action),
+		Debug:             selection.Debug,
+	}
+}
